Moved ArmInterface into the new arm_interface module
